Read temperature, use random values for ph & spin

+++ uno_slave/spin.cpp
@@ -1,8 +1,9 @@
+#include <Arduino.h>
 #include "spin.h"
 void spinSetup() {
 
 }
 
-float spinLoop() {
+unsigned long spinLoop() {
   return (float)random(1000)/10.0 + 400.0; 
 }
